Improve debouncing logic to make the ui snappier
* If the render is fast enough, render real-time
* If the render does not take too long, just debounce highly interactive
  motions (e.g. dragging)
* If the render is very slow, debounce all motions to allow for multiple
  changes to be made by the user before re-rendering

This also includes making the re-render pass logic more abstract and
tracking the time between image updates for estimating debounce time.

+++ src/image_gen/probe.rs
@@ -27,7 +27,7 @@ impl FromFloat for f32 {
 /// The resulting vector will be of length `max_iter` or less if the point escapes.
 pub fn probe<T>(
     ProbeLocation { x, y }: &ProbeLocation,
-    max_iter: usize,
+    max_iter: u64,
     zoom: f64,
 ) -> (Vec<[T; 2]>, Vec<[T; 2]>)
 where
